@@ -1,6 +1,6 @@
 module github.com/findyourpaths/goskyr
 
-go 1.23.10
+go 1.25.0
 
 require (
 	github.com/PuerkitoBio/goquery v1.10.2
